Extract codex-extensions from codex-core
