+++ codex-rs/Cargo.toml
@@ -27,6 +27,7 @@ members = [
     "exec",
     "execpolicy",
     "execpolicy-legacy",
+    "extensions",
     "keyring-store",
     "file-search",
     "linux-sandbox",
@@ -105,6 +106,7 @@ codex-config = { path = "config" }
 codex-core = { path = "core" }
 codex-exec = { path = "exec" }
 codex-execpolicy = { path = "execpolicy" }
+codex-extensions = { path = "extensions" }
 codex-experimental-api-macros = { path = "codex-experimental-api-macros" }
 codex-feedback = { path = "feedback" }
 codex-file-search = { path = "file-search" }
