+++ codex-rs/core/src/skills/mod.rs
@@ -15,9 +15,9 @@ pub(crate) use injection::build_skill_injections;
 pub(crate) use injection::collect_explicit_skill_mentions;
 pub(crate) use invocation_utils::build_implicit_skill_path_indexes;
 pub(crate) use invocation_utils::maybe_emit_implicit_skill_invocation;
+pub use codex_extensions::skills::SkillError;
+pub use codex_extensions::skills::SkillLoadOutcome;
+pub use codex_extensions::skills::SkillMetadata;
+pub use codex_extensions::skills::SkillPolicy;
 pub use manager::SkillsManager;
-pub use model::SkillError;
-pub use model::SkillLoadOutcome;
-pub use model::SkillMetadata;
-pub use model::SkillPolicy;
 pub use render::render_skills_section;
