+++ codex-rs/extensions/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod plugins;
+pub mod skills;
