+++ codex-rs/extensions/src/skills/mod.rs
@@ -1,0 +1,12 @@
+pub mod model;
+mod render;
+
+pub use model::SkillDependencies;
+pub use model::SkillError;
+pub use model::SkillInterface;
+pub use model::SkillLoadOutcome;
+pub use model::SkillManagedNetworkOverride;
+pub use model::SkillMetadata;
+pub use model::SkillPolicy;
+pub use model::SkillToolDependency;
+pub use render::render_skills_section;
